ci: adopt org reusable workflows and pass their gates (#78)
Adopt the org reusable CI (go-ci/dco/line-limit/main-guard callers, GitHub/Glyndor-owned actions only, SHA-pinned) and bring the repo to pass the stricter gates it introduces:

- CI migration: thin callers, gosec/govulncheck run directly, scheduled fuzz, drop Codecov/Go Report Card/OpenSSF Scorecard/semantic-PR-title; dependabot targets develop
- make the bundled examples standalone modules so they leave the root coverage scope (total 90.0%)
- split jwt_test.go (836 -> 327/283/244 code lines) under the 500-line limit
- bump x/net to v0.56.0 and the go directive to 1.26.3 -> govulncheck clean

Closes #77.
