+++ .github/dependabot.yml
@@ -4,14 +4,17 @@ updates:
     directory: "/"
     schedule:
       interval: "daily"
+    target-branch: "develop"
     labels:
-      - "dependencies"
-      - "go"
+      - "type:deps"
+    open-pull-requests-limit: 10
 
   - package-ecosystem: "github-actions"
     directory: "/"
     schedule:
       interval: "daily"
+    target-branch: "develop"
     labels:
-      - "dependencies"
-      - "github-actions"
+      - "type:deps"
+      - "type:ci"
+    open-pull-requests-limit: 10
