+++ .github/workflows/dco.yml
@@ -1,0 +1,14 @@
+name: DCO
+
+on:
+  pull_request:
+    branches:
+      - main
+      - develop
+
+permissions:
+  contents: read
+
+jobs:
+  dco:
+    uses: Glyndor/.github/.github/workflows/dco.yml@80523422631014f591b07bc24c507145faabb271 # v1.2.0
