+++ .github/workflows/fuzz.yml
@@ -1,0 +1,34 @@
+name: Fuzz
+
+on:
+  schedule:
+    - cron: "0 7 * * 1"  # Every Monday at 07:00 UTC
+  workflow_dispatch:
+
+permissions:
+  contents: read
+
+jobs:
+  fuzz:
+    name: go test -fuzz (smoke)
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        include:
+          - { package: ./auth/email, func: FuzzValidateAndNormalize }
+          - { package: ./auth/username, func: FuzzValidateAndNormalize }
+          - { package: ./auth/jwt, func: FuzzVerifyAccessToken }
+          - { package: ./auth/jwt, func: FuzzIsUUIDv7 }
+          - { package: ./auth/password, func: FuzzValidatePolicy }
+          - { package: ./auth/password, func: FuzzParsePHC }
+    steps:
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+
+      - uses: actions/setup-go@40f1582b2485089dde7abd97c1529aa768e1baff # v5
+        with:
+          go-version: '1.26.1'
+          cache: true
+
+      - name: Fuzz ${{ matrix.func }} (${{ matrix.package }})
+        run: go test -run='^$' -fuzz='^${{ matrix.func }}$' -fuzztime=60s ${{ matrix.package }}
