+++ .github/workflows/govulncheck.yml
@@ -2,9 +2,9 @@ name: Govulncheck
 
 on:
   push:
-    branches: [ main ]
+    branches: [main, develop]
   pull_request:
-    branches: [ main ]
+    branches: [main, develop]
   schedule:
     - cron: '0 6 * * 1'
 
@@ -24,8 +24,8 @@ jobs:
           cache: true
 
       - name: Install govulncheck
-        # Pin to a specific tag instead of @latest so scorecard's
-        # PinnedDependenciesID check passes and the workflow is reproducible.
+        # Pin to a specific tag instead of @latest so the workflow is
+        # reproducible and the dependency stays SHA/version-pinned.
         run: go install golang.org/x/vuln/cmd/govulncheck@v1.1.4
 
       - name: Run govulncheck
