+++ .github/workflows/line-limit.yml
@@ -1,0 +1,14 @@
+name: Line limit
+
+on:
+  pull_request:
+    branches:
+      - main
+      - develop
+
+permissions:
+  contents: read
+
+jobs:
+  line-limit:
+    uses: Glyndor/.github/.github/workflows/line-limit.yml@80523422631014f591b07bc24c507145faabb271 # v1.2.0
