+++ .github/workflows/main-guard.yml
@@ -1,0 +1,11 @@
+name: guard
+
+on:
+  pull_request:
+    branches: ["main"]
+
+permissions: {}
+
+jobs:
+  develop-only:
+    uses: Glyndor/.github/.github/workflows/main-guard.yml@80523422631014f591b07bc24c507145faabb271 # v1.2.0
