+++ auth/jwt/jwt_test.go
@@ -14,8 +14,6 @@ import (
 	"testing"
 	"time"
 
-	gjwt "github.com/golang-jwt/jwt/v5"
-
 	"github.com/Jaro-c/authcore"
 	"github.com/Jaro-c/authcore/internal/clock"
 )
@@ -406,673 +404,5 @@ func TestCreateTokens_wrongAudienceRejectsToken(t *testing.T) {
 	_, err := j2.VerifyAccessToken(pair.AccessToken)
 	if err == nil {
 		t.Error("expected error when audience does not match, got nil")
-	}
-}
-
-func TestVerifyAccessToken_wrongIssuerRejectsToken(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.Issuer = "https://auth.service-a.example.com"
-	j := newTestJWT[struct{}](t, newFakeProvider(t), cfg)
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	// Another service that happens to share the signing key (e.g. accidental
-	// key reuse across microservices) must not accept a token issued by the
-	// first service. The iss claim must be checked on verification.
-	cfg2 := DefaultConfig()
-	cfg2.Issuer = "https://auth.service-b.example.com"
-	j2 := newTestJWT[struct{}](t, newFakeProvider(t), cfg2)
-	j2.priv = j.priv
-	j2.pub = j.pub
-
-	_, err := j2.VerifyAccessToken(pair.AccessToken)
-	if !errors.Is(err, ErrTokenInvalid) {
-		t.Errorf("expected ErrTokenInvalid when issuer does not match, got %v", err)
-	}
-}
-
-func TestVerifyAccessToken_missingKidRejectsToken(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-
-	// An attacker may try to strip the kid header entirely, betting that
-	// the verifier trusts the (single) configured public key regardless.
-	// The verifier must reject tokens whose kid does not match, which
-	// includes tokens where kid is absent.
-	claims := newAccessClaims(j.cfg.Issuer, testSubject, "019600ab-0000-7000-8000-000000000002", j.cfg.Audience, struct{}{}, epoch, j.cfg.AccessTokenTTL)
-	claims.Type = tokenTypeAccess
-	token := gjwt.NewWithClaims(gjwt.SigningMethodEdDSA, claims)
-	// Deliberately do not set token.Header["kid"].
-	stripped, err := token.SignedString(j.priv)
-	if err != nil {
-		t.Fatalf("sign kid-less token: %v", err)
-	}
-
-	_, err = j.VerifyAccessToken(stripped)
-	if !errors.Is(err, ErrTokenInvalid) {
-		t.Errorf("expected ErrTokenInvalid for missing kid, got %v", err)
-	}
-}
-
-func TestRotateTokens_unknownKidRejectsRefreshToken(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-
-	// Mirror of the access-token kid check but on the rotation path. A
-	// refresh token with a kid the module does not recognise must be
-	// rejected before it is exchanged for a fresh pair.
-	claims := newRefreshClaims(j.cfg.Issuer, testSubject, "019600ab-0000-7000-8000-000000000003", j.cfg.Audience, epoch, j.cfg.RefreshTokenTTL)
-	claims.Type = tokenTypeRefresh
-	token := gjwt.NewWithClaims(gjwt.SigningMethodEdDSA, claims)
-	token.Header["kid"] = "attacker-controlled-kid"
-	tampered, err := token.SignedString(j.priv)
-	if err != nil {
-		t.Fatalf("sign tampered refresh token: %v", err)
-	}
-
-	_, err = j.RotateTokens(tampered, struct{}{})
-	if !errors.Is(err, ErrTokenInvalid) {
-		t.Errorf("expected ErrTokenInvalid for unknown kid on refresh, got %v", err)
-	}
-}
-
-func TestVerifyAccessToken_unknownKidRejectsToken(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-
-	// Craft a token signed with the module's private key (so the signature
-	// is valid) but with a "kid" header the verifier does not recognise.
-	// This simulates an attacker who has obtained a signing key but whose
-	// "kid" is not on the verifier's trusted list — the kid check must
-	// reject the token before it is accepted.
-	claims := newAccessClaims(j.cfg.Issuer, testSubject, "019600ab-0000-7000-8000-000000000001", j.cfg.Audience, struct{}{}, epoch, j.cfg.AccessTokenTTL)
-	claims.Type = tokenTypeAccess
-	token := gjwt.NewWithClaims(gjwt.SigningMethodEdDSA, claims)
-	token.Header["kid"] = "attacker-controlled-kid"
-	tampered, err := token.SignedString(j.priv)
-	if err != nil {
-		t.Fatalf("sign tampered token: %v", err)
-	}
-
-	_, err = j.VerifyAccessToken(tampered)
-	if !errors.Is(err, ErrTokenInvalid) {
-		t.Errorf("expected ErrTokenInvalid for unknown kid, got %v", err)
-	}
-}
-
-func TestRotateTokens_wrongIssuerRejectsRefreshToken(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.Issuer = "https://auth.service-a.example.com"
-	j := newTestJWT[struct{}](t, newFakeProvider(t), cfg)
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	cfg2 := DefaultConfig()
-	cfg2.Issuer = "https://auth.service-b.example.com"
-	j2 := newTestJWT[struct{}](t, newFakeProvider(t), cfg2)
-	j2.priv = j.priv
-	j2.pub = j.pub
-
-	_, err := j2.RotateTokens(pair.RefreshToken, struct{}{})
-	if !errors.Is(err, ErrTokenInvalid) {
-		t.Errorf("expected ErrTokenInvalid when issuer does not match, got %v", err)
-	}
-}
-
-func TestRotateTokens_audienceEmbeddedInRefreshToken(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.Audience = []string{"https://api.example.com"}
-	j := newTestJWT[struct{}](t, newFakeProvider(t), cfg)
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	// Rotation must succeed — refresh token carries the same audience.
-	j.clock = clock.Fixed(epoch.Add(time.Second))
-	_, err := j.RotateTokens(pair.RefreshToken, struct{}{})
-	if err != nil {
-		t.Fatalf("RotateTokens() error = %v", err)
-	}
-}
-
-// ---- custom claims ----------------------------------------------------------
-
-type testUserClaims struct {
-	Name string `json:"name"`
-	Role string `json:"role"`
-}
-
-func TestCreateTokens_customClaimsRoundTrip(t *testing.T) {
-	j := newTestJWT[testUserClaims](t, newFakeProvider(t), DefaultConfig())
-	extra := testUserClaims{Name: "Juan", Role: "admin"}
-
-	pair, err := j.CreateTokens(testSubject, extra)
-	if err != nil {
-		t.Fatalf("CreateTokens() error = %v", err)
-	}
-
-	claims, err := j.VerifyAccessToken(pair.AccessToken)
-	if err != nil {
-		t.Fatalf("VerifyAccessToken() error = %v", err)
-	}
-	if claims.Extra.Name != "Juan" {
-		t.Errorf("Extra.Name = %q, want %q", claims.Extra.Name, "Juan")
-	}
-	if claims.Extra.Role != "admin" {
-		t.Errorf("Extra.Role = %q, want %q", claims.Extra.Role, "admin")
-	}
-}
-
-func TestCreateTokens_refreshTokenHasNoExtraClaims(t *testing.T) {
-	j := newTestJWT[testUserClaims](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, testUserClaims{Name: "Juan", Role: "admin"})
-
-	// Decode the refresh token payload and verify "extra" is absent.
-	parts := strings.SplitN(pair.RefreshToken, ".", 3)
-	data, _ := base64.RawURLEncoding.DecodeString(parts[1])
-	var payload map[string]any
-	if err := json.Unmarshal(data, &payload); err != nil {
-		t.Fatalf("unmarshal refresh payload: %v", err)
-	}
-	if _, ok := payload["extra"]; ok {
-		t.Error("refresh token payload must not contain 'extra' field")
-	}
-}
-
-func TestCreateTokens_refreshTokenHasIat(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	parts := strings.SplitN(pair.RefreshToken, ".", 3)
-	data, _ := base64.RawURLEncoding.DecodeString(parts[1])
-	var payload map[string]any
-	if err := json.Unmarshal(data, &payload); err != nil {
-		t.Fatalf("unmarshal refresh payload: %v", err)
-	}
-	if _, ok := payload["iat"]; !ok {
-		t.Error("refresh token payload must contain 'iat' field")
-	}
-}
-
-// ---- kid header -------------------------------------------------------------
-
-func TestCreateTokens_accessTokenHeaderContainsKid(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	h := tokenHeader(t, pair.AccessToken)
-	kid, ok := h["kid"].(string)
-	if !ok || kid == "" {
-		t.Errorf("access token header missing or empty kid: %v", h)
-	}
-	if kid != "test0000test0000" {
-		t.Errorf("kid = %q, want %q", kid, "test0000test0000")
-	}
-}
-
-func TestCreateTokens_refreshTokenHeaderContainsKid(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	h := tokenHeader(t, pair.RefreshToken)
-	kid, ok := h["kid"].(string)
-	if !ok || kid == "" {
-		t.Errorf("refresh token header missing or empty kid: %v", h)
-	}
-	if kid != "test0000test0000" {
-		t.Errorf("kid = %q, want %q", kid, "test0000test0000")
-	}
-}
-
-func TestCreateTokens_bothTokensShareSameKid(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	accessKid := tokenHeader(t, pair.AccessToken)["kid"]
-	refreshKid := tokenHeader(t, pair.RefreshToken)["kid"]
-	if accessKid != refreshKid {
-		t.Errorf("access kid %q != refresh kid %q", accessKid, refreshKid)
-	}
-}
-
-// ---- access token jti -------------------------------------------------------
-
-func TestCreateTokens_sessionIDIsUUIDv7Format(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	id := pair.SessionID
-	if len(id) != 36 {
-		t.Fatalf("SessionID length = %d, want 36", len(id))
-	}
-	if id[14] != '7' {
-		t.Errorf("SessionID version digit = %q, want '7'", id[14])
-	}
-	if id[19] != '8' && id[19] != '9' && id[19] != 'a' && id[19] != 'b' {
-		t.Errorf("SessionID variant nibble = %q, want 8/9/a/b", id[19])
-	}
-}
-
-func TestCreateTokens_accessAndRefreshShareSameJTI(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	claims, err := j.VerifyAccessToken(pair.AccessToken)
-	if err != nil {
-		t.Fatalf("VerifyAccessToken() error = %v", err)
-	}
-	if claims.TokenID != pair.SessionID {
-		t.Errorf("access token jti %q must equal SessionID %q", claims.TokenID, pair.SessionID)
-	}
-}
-
-// ---- VerifyAccessToken() ----------------------------------------------------
-
-func TestVerifyAccessToken_validTokenReturnsCorrectClaims(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens("0191b432-b5a7-7c4f-b2e6-7a3f1d2e0099", struct{}{})
-
-	claims, err := j.VerifyAccessToken(pair.AccessToken)
-	if err != nil {
-		t.Fatalf("VerifyAccessToken() error = %v", err)
-	}
-	if claims.Subject != "0191b432-b5a7-7c4f-b2e6-7a3f1d2e0099" {
-		t.Errorf("Subject = %q, want %q", claims.Subject, "0191b432-b5a7-7c4f-b2e6-7a3f1d2e0099")
-	}
-	if !claims.IssuedAt.Equal(epoch) {
-		t.Errorf("IssuedAt = %v, want %v", claims.IssuedAt, epoch)
-	}
-}
-
-func TestVerifyAccessToken_expiredTokenReturnsErrTokenExpired(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.AccessTokenTTL = 10 * time.Minute
-	j := newTestJWT[struct{}](t, newFakeProvider(t), cfg)
-
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	// Advance clock past expiry.
-	j.clock = clock.Fixed(epoch.Add(11 * time.Minute))
-
-	_, err := j.VerifyAccessToken(pair.AccessToken)
-	if !errors.Is(err, ErrTokenExpired) {
-		t.Errorf("expected ErrTokenExpired, got %v", err)
-	}
-}
-
-func TestVerifyAccessToken_tokenAtExactExpiryIsExpired(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.AccessTokenTTL = 10 * time.Minute
-	j := newTestJWT[struct{}](t, newFakeProvider(t), cfg)
-
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-	// golang-jwt/v5 uses strict now.After(exp), so the token is still valid at
-	// the exact expiry second. Advance one second past exp to trigger expiry.
-	j.clock = clock.Fixed(epoch.Add(10*time.Minute + time.Second))
-
-	_, err := j.VerifyAccessToken(pair.AccessToken)
-	if !errors.Is(err, ErrTokenExpired) {
-		t.Errorf("token one second past expiry should return ErrTokenExpired, got %v", err)
-	}
-}
-
-func TestVerifyAccessToken_tamperedSignatureReturnsErrTokenInvalid(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	// Modify a middle character of the signature segment to guarantee the
-	// decoded bytes change. The last base64url char of an Ed25519 signature
-	// encodes only 2 meaningful bits — altering it does not affect the
-	// decoded byte slice, so we target a position well inside the signature.
-	token := pair.AccessToken
-	mid := len(token) - 10
-	orig := token[mid]
-	replacement := byte('A')
-	if orig == 'A' {
-		replacement = 'B'
-	}
-	tampered := token[:mid] + string(replacement) + token[mid+1:]
-
-	_, err := j.VerifyAccessToken(tampered)
-	if !errors.Is(err, ErrTokenInvalid) && !errors.Is(err, ErrTokenMalformed) {
-		t.Errorf("expected ErrTokenInvalid or ErrTokenMalformed, got %v", err)
-	}
-}
-
-func TestVerifyAccessToken_malformedTokenReturnsErrTokenMalformed(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-
-	cases := []string{"", "only-one-part", "two.parts", "a.b.c.d"}
-	for _, tc := range cases {
-		_, err := j.VerifyAccessToken(tc)
-		if !errors.Is(err, ErrTokenMalformed) && !errors.Is(err, ErrTokenInvalid) {
-			t.Errorf("input %q: expected ErrTokenMalformed, got %v", tc, err)
-		}
-	}
-}
-
-func TestVerifyAccessToken_rejectsRefreshToken(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	_, err := j.VerifyAccessToken(pair.RefreshToken)
-	if !errors.Is(err, ErrWrongTokenType) {
-		t.Errorf("expected ErrWrongTokenType, got %v", err)
-	}
-}
-
-func TestVerifyAccessToken_rejectsTokenFromDifferentKey(t *testing.T) {
-	p1 := newFakeProvider(t)
-	p2 := newFakeProvider(t) // different Ed25519 key pair
-
-	j1 := newTestJWT[struct{}](t, p1, DefaultConfig())
-	j2 := newTestJWT[struct{}](t, p2, DefaultConfig())
-
-	pair, _ := j1.CreateTokens(testSubject, struct{}{})
-
-	_, err := j2.VerifyAccessToken(pair.AccessToken)
-	if !errors.Is(err, ErrTokenInvalid) {
-		t.Errorf("token signed by j1 should be invalid for j2, got %v", err)
-	}
-}
-
-// ---- HashRefreshToken() -----------------------------------------------------
-
-func TestHashRefreshToken_isDeterministic(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	h1 := j.HashRefreshToken(pair.RefreshToken)
-	h2 := j.HashRefreshToken(pair.RefreshToken)
-	if h1 != h2 {
-		t.Errorf("HashRefreshToken is not deterministic: %q != %q", h1, h2)
-	}
-}
-
-func TestHashRefreshToken_differentTokensDifferentHashes(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	p1, _ := j.CreateTokens(testSubject, struct{}{})
-	p2, _ := j.CreateTokens(testSubject, struct{}{})
-
-	if j.HashRefreshToken(p1.RefreshToken) == j.HashRefreshToken(p2.RefreshToken) {
-		t.Error("different refresh tokens must produce different HMAC digests")
-	}
-}
-
-func TestHashRefreshToken_differentSecretsDifferentHashes(t *testing.T) {
-	j1 := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	j2 := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-
-	// Sign a token with j1's key, hash it with both modules' secrets.
-	pair, _ := j1.CreateTokens(testSubject, struct{}{})
-
-	h1 := j1.HashRefreshToken(pair.RefreshToken)
-	h2 := j2.HashRefreshToken(pair.RefreshToken)
-
-	// Different HMAC secrets must produce different digests.
-	if h1 == h2 {
-		t.Error("same token hashed with different secrets should produce different results")
-	}
-}
-
-// ---- RotateTokens() ---------------------------------------------------------
-
-func TestRotateTokens_returnsNewPairForSameSubject(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	old, _ := j.CreateTokens("0191b432-b5a7-7c4f-b2e6-7a3f1d2e0007", struct{}{})
-
-	newPair, err := j.RotateTokens(old.RefreshToken, struct{}{})
-	if err != nil {
-		t.Fatalf("RotateTokens() error = %v", err)
-	}
-
-	claims, _ := j.VerifyAccessToken(newPair.AccessToken)
-	if claims.Subject != "0191b432-b5a7-7c4f-b2e6-7a3f1d2e0007" {
-		t.Errorf("Subject after rotation = %q, want %q", claims.Subject, "0191b432-b5a7-7c4f-b2e6-7a3f1d2e0007")
-	}
-}
-
-func TestRotateTokens_newTokensDifferFromOld(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	old, _ := j.CreateTokens(testSubject, struct{}{})
-
-	// Advance clock by 1 second so the issued-at differs.
-	j.clock = clock.Fixed(epoch.Add(time.Second))
-	newPair, _ := j.RotateTokens(old.RefreshToken, struct{}{})
-
-	if newPair.AccessToken == old.AccessToken {
-		t.Error("new AccessToken must differ from old AccessToken")
-	}
-	if newPair.RefreshToken == old.RefreshToken {
-		t.Error("new RefreshToken must differ from old RefreshToken")
-	}
-	if newPair.RefreshTokenHash == old.RefreshTokenHash {
-		t.Error("new RefreshTokenHash must differ from old RefreshTokenHash")
-	}
-}
-
-func TestRotateTokens_expiredRefreshTokenReturnsErrTokenExpired(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.RefreshTokenTTL = 24 * time.Hour
-	j := newTestJWT[struct{}](t, newFakeProvider(t), cfg)
-
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	// Advance past refresh token TTL.
-	j.clock = clock.Fixed(epoch.Add(25 * time.Hour))
-
-	_, err := j.RotateTokens(pair.RefreshToken, struct{}{})
-	if !errors.Is(err, ErrTokenExpired) {
-		t.Errorf("expected ErrTokenExpired, got %v", err)
-	}
-}
-
-func TestRotateTokens_rejectsAccessToken(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	_, err := j.RotateTokens(pair.AccessToken, struct{}{})
-	if !errors.Is(err, ErrWrongTokenType) {
-		t.Errorf("expected ErrWrongTokenType, got %v", err)
-	}
-}
-
-func TestRotateTokens_rejectsTamperedToken(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	rt := pair.RefreshToken
-	midRT := len(rt) - 10
-	origRT := rt[midRT]
-	replacementRT := byte('A')
-	if origRT == 'A' {
-		replacementRT = 'B'
-	}
-	tampered := rt[:midRT] + string(replacementRT) + rt[midRT+1:]
-	_, err := j.RotateTokens(tampered, struct{}{})
-	if !errors.Is(err, ErrTokenInvalid) && !errors.Is(err, ErrTokenMalformed) {
-		t.Errorf("expected ErrTokenInvalid or ErrTokenMalformed, got %v", err)
-	}
-}
-
-func TestRotateTokens_newHashMatchesHashRefreshToken(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	j.clock = clock.Fixed(epoch.Add(time.Second))
-	newPair, _ := j.RotateTokens(pair.RefreshToken, struct{}{})
-
-	got := j.HashRefreshToken(newPair.RefreshToken)
-	if got != newPair.RefreshTokenHash {
-		t.Errorf("HashRefreshToken(new token) = %q, want %q", got, newPair.RefreshTokenHash)
-	}
-}
-
-func TestRotateTokens_preservesSessionID(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	original, _ := j.CreateTokens(testSubject, struct{}{})
-
-	j.clock = clock.Fixed(epoch.Add(time.Second))
-	rotated, err := j.RotateTokens(original.RefreshToken, struct{}{})
-	if err != nil {
-		t.Fatalf("RotateTokens() error = %v", err)
-	}
-	if rotated.SessionID != original.SessionID {
-		t.Errorf("SessionID changed after rotation: got %q, want %q", rotated.SessionID, original.SessionID)
-	}
-}
-
-// ---- VerifyRefreshTokenHash() -----------------------------------------------
-
-func TestVerifyRefreshTokenHash_matchesStoredHash(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	if !j.VerifyRefreshTokenHash(pair.RefreshToken, pair.RefreshTokenHash) {
-		t.Error("VerifyRefreshTokenHash returned false for a valid token/hash pair")
-	}
-}
-
-func TestVerifyRefreshTokenHash_modifiedTokenDoesNotMatch(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	if j.VerifyRefreshTokenHash(pair.RefreshToken+"X", pair.RefreshTokenHash) {
-		t.Error("VerifyRefreshTokenHash returned true for a modified token")
-	}
-}
-
-func TestVerifyRefreshTokenHash_modifiedHashDoesNotMatch(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	badHash := strings.Repeat("0", 64)
-	if j.VerifyRefreshTokenHash(pair.RefreshToken, badHash) {
-		t.Error("VerifyRefreshTokenHash returned true for a tampered hash")
-	}
-}
-
-func TestVerifyRefreshTokenHash_emptyTokenDoesNotMatch(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	if j.VerifyRefreshTokenHash("", pair.RefreshTokenHash) {
-		t.Error("VerifyRefreshTokenHash returned true for empty token")
-	}
-}
-
-func TestVerifyRefreshTokenHash_emptyHashDoesNotMatch(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	if j.VerifyRefreshTokenHash(pair.RefreshToken, "") {
-		t.Error("VerifyRefreshTokenHash returned true for empty stored hash")
-	}
-}
-
-func TestVerifyRefreshTokenHash_consistentWithHashRefreshToken(t *testing.T) {
-	j := newTestJWT[struct{}](t, newFakeProvider(t), DefaultConfig())
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	hash := j.HashRefreshToken(pair.RefreshToken)
-	if !j.VerifyRefreshTokenHash(pair.RefreshToken, hash) {
-		t.Error("VerifyRefreshTokenHash disagrees with HashRefreshToken for the same input")
-	}
-}
-
-// ---- ClockSkewLeeway --------------------------------------------------------
-
-func TestNew_negativeLeewayReturnsErrInvalidConfig(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.ClockSkewLeeway = -1 * time.Second
-
-	_, err := New[struct{}](newFakeProvider(t), cfg)
-	if !errors.Is(err, ErrInvalidConfig) {
-		t.Errorf("expected ErrInvalidConfig, got %v", err)
-	}
-}
-
-func TestVerifyAccessToken_tokenWithinLeewayIsAccepted(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.AccessTokenTTL = 10 * time.Minute
-	cfg.ClockSkewLeeway = 30 * time.Second
-	j := newTestJWT[struct{}](t, newFakeProvider(t), cfg)
-
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	// Advance 20 s past expiry — within the 30 s leeway window.
-	j.clock = clock.Fixed(epoch.Add(10*time.Minute + 20*time.Second))
-	if _, err := j.VerifyAccessToken(pair.AccessToken); err != nil {
-		t.Errorf("token within leeway should be accepted, got %v", err)
-	}
-}
-
-func TestVerifyAccessToken_tokenBeyondLeewayIsRejected(t *testing.T) {
-	cfg := DefaultConfig()
-	cfg.AccessTokenTTL = 10 * time.Minute
-	cfg.ClockSkewLeeway = 30 * time.Second
-	j := newTestJWT[struct{}](t, newFakeProvider(t), cfg)
-
-	pair, _ := j.CreateTokens(testSubject, struct{}{})
-
-	// Advance 31 s past expiry — beyond the leeway window.
-	j.clock = clock.Fixed(epoch.Add(10*time.Minute + 31*time.Second))
-	_, err := j.VerifyAccessToken(pair.AccessToken)
-	if !errors.Is(err, ErrTokenExpired) {
-		t.Errorf("token beyond leeway should return ErrTokenExpired, got %v", err)
-	}
-}
-
-// ---- applyDefaults ----------------------------------------------------------
-
-func TestApplyDefaults_zeroTTLsAreFilledFromDefaults(t *testing.T) {
-	cfg := Config{
-		Issuer:   "my-service",
-		Audience: []string{"my-audience"},
-		// AccessTokenTTL and RefreshTokenTTL intentionally zero
-	}
-	j, err := New[struct{}](newFakeProvider(t), cfg)
-	if err != nil {
-		t.Fatalf("New() with zero TTLs error = %v", err)
-	}
-	if j.cfg.AccessTokenTTL != DefaultConfig().AccessTokenTTL {
-		t.Errorf("AccessTokenTTL not filled: got %v", j.cfg.AccessTokenTTL)
-	}
-	if j.cfg.RefreshTokenTTL != DefaultConfig().RefreshTokenTTL {
-		t.Errorf("RefreshTokenTTL not filled: got %v", j.cfg.RefreshTokenTTL)
-	}
-}
-
-func TestApplyDefaults_zeroIssuerIsFilledFromDefaults(t *testing.T) {
-	cfg := Config{Audience: []string{"x"}}
-	j, err := New[struct{}](newFakeProvider(t), cfg)
-	if err != nil {
-		t.Fatalf("New() error = %v", err)
-	}
-	if j.cfg.Issuer != DefaultConfig().Issuer {
-		t.Errorf("Issuer not filled: got %q", j.cfg.Issuer)
-	}
-}
-
-func TestApplyDefaults_zeroAudienceIsFilledFromDefaults(t *testing.T) {
-	cfg := Config{Issuer: "x"}
-	j, err := New[struct{}](newFakeProvider(t), cfg)
-	if err != nil {
-		t.Fatalf("New() error = %v", err)
-	}
-	if len(j.cfg.Audience) == 0 {
-		t.Error("Audience not filled from defaults")
-	}
-}
-
-// ---- TestRotateTokens_freshClaimsEmbeddedInNewAccessToken -------------------
-
-func TestRotateTokens_freshClaimsEmbeddedInNewAccessToken(t *testing.T) {
-	j := newTestJWT[testUserClaims](t, newFakeProvider(t), DefaultConfig())
-	old, _ := j.CreateTokens(testSubject, testUserClaims{Name: "Juan", Role: "user"})
-
-	j.clock = clock.Fixed(epoch.Add(time.Second))
-	fresh := testUserClaims{Name: "Juan", Role: "admin"} // role promoted
-	newPair, _ := j.RotateTokens(old.RefreshToken, fresh)
-
-	claims, err := j.VerifyAccessToken(newPair.AccessToken)
-	if err != nil {
-		t.Fatalf("VerifyAccessToken() error = %v", err)
-	}
-	if claims.Extra.Role != "admin" {
-		t.Errorf("Extra.Role after rotation = %q, want %q", claims.Extra.Role, "admin")
 	}
 }
