--- codecov.yml
@@ -1,27 +1,0 @@
-# Codecov configuration for authcore.
-# See https://docs.codecov.com/docs/codecov-yaml for the full schema.
-
-coverage:
-  status:
-    project:
-      default:
-        target: auto        # never regress overall coverage
-        threshold: 1%       # tolerate small fluctuations
-    patch:
-      default:
-        target: 80%         # PRs must keep new code at 80%+ coverage
-        threshold: 5%       # allow small dips under target
-        only_pulls: true
-
-# Files and paths excluded from coverage analysis.
-# Documentation and example programs are not test targets.
-ignore:
-  - "examples/**"
-  - "**/example_test.go"
-  - "**/*_bench_test.go"
-  - "**/*_fuzz_test.go"
-
-comment:
-  layout: "reach,diff,flags,files"
-  behavior: default
-  require_changes: false
