+++ examples/basic/go.mod
@@ -1,0 +1,7 @@
+module github.com/Jaro-c/authcore/examples/basic
+
+go 1.26.2
+
+require github.com/Jaro-c/authcore v1.2.2
+
+replace github.com/Jaro-c/authcore => ../../
