+++ examples/jwt/go.mod
@@ -1,0 +1,9 @@
+module github.com/Jaro-c/authcore/examples/jwt
+
+go 1.26.2
+
+require github.com/Jaro-c/authcore v1.2.2
+
+require github.com/golang-jwt/jwt/v5 v5.3.1 // indirect
+
+replace github.com/Jaro-c/authcore => ../../
