+++ examples/password/go.mod
@@ -1,0 +1,13 @@
+module github.com/Jaro-c/authcore/examples/password
+
+go 1.26.2
+
+require github.com/Jaro-c/authcore v1.2.2
+
+require (
+	golang.org/x/crypto v0.53.0 // indirect
+	golang.org/x/sys v0.46.0 // indirect
+	golang.org/x/text v0.38.0 // indirect
+)
+
+replace github.com/Jaro-c/authcore => ../../
