+++ go.mod
@@ -1,13 +1,13 @@
 module github.com/Jaro-c/authcore
 
-go 1.26.1
+go 1.26.3
 
 require (
 	github.com/golang-jwt/jwt/v5 v5.3.1
-	golang.org/x/crypto v0.50.0
-	golang.org/x/net v0.53.0
-	golang.org/x/sync v0.20.0
-	golang.org/x/text v0.36.0
+	golang.org/x/crypto v0.53.0
+	golang.org/x/net v0.56.0
+	golang.org/x/sync v0.21.0
+	golang.org/x/text v0.38.0
 )
 
-require golang.org/x/sys v0.43.0 // indirect
+require golang.org/x/sys v0.46.0 // indirect
